+++ main.rb
@@ -1,6 +1,5 @@
 require './app'
 
-
 def main
   app = App.new
   app.start_app
